refactor & modularize init
